It's working mostly, I still have a slight bug...
ADD() calls SUB(), which calls ADD() again, and so on, until the stack blows up, if the sign of A (the lhs operand) is neg ('-'). I have to do some manual work, because watching a GCSE video on Youtube is a bit too humiliating. I will truth-table tc outta this, mark my words!!! (No, I'll just do some trusty paper-based testing).

P.S. The "colorconsole" dependency is available here on my Github profile. It is currently needed for "make aplib" to compile.