Update, I keep breaking it.
The DIV (DIVIDE) operator is almost working; there are edge-cases where it presently fails. A number can be divided by a larger number (divisor) as long as the larger number isn't a repeat of the digit(s) in the dividend.  Also, during these edge case ranges, a too-long 'B' value (divisor) causes the function to crash.